# FFM-11714 - More memory optimisations for response encoding