**What**

- Uses json.NewEncoder for encoding the response body

**Why**

- Benchmarks show this is pretty much as fast as jsoniter.Marshal but
  uses less memory and fewer allocations

**Testing**

- Ran benchmarks
```
Benchmark_EncodeResponse
Benchmark_EncodeResponse/jsoniter.Marshal
Benchmark_EncodeResponse/jsoniter.Marshal-12         	    4872	    227514 ns/op	  155079 B/op	      12 allocs/op
Benchmark_EncodeResponse/json.Marshal
Benchmark_EncodeResponse/json.Marshal-12             	    4597	    260348 ns/op	  151189 B/op	      12 allocs/op
Benchmark_EncodeResponse/json.NewEncoder
Benchmark_EncodeResponse/json.NewEncoder-12          	    4885	    236368 ns/op	   75636 B/op	      11 allocs/op
Benchmark_EncodeResponse/jsoniter.NewEncoder
Benchmark_EncodeResponse/jsoniter.NewEncoder-12      	    3776	    293798 ns/op	  359755 B/op	      29 allocs/op
```